+++ Jellyfin.Xtream/PluginServiceRegistrator.cs
@@ -15,6 +15,7 @@
 
 using Jellyfin.Xtream.Client;
 using Jellyfin.Xtream.Providers;
+using Jellyfin.Xtream.Service;
 using Jellyfin.Xtream.Tasks;
 using MediaBrowser.Controller;
 using MediaBrowser.Controller.Channels;
@@ -33,6 +34,8 @@ public class PluginServiceRegistrator : IPluginServiceRegistrator
     public void RegisterServices(IServiceCollection serviceCollection, IServerApplicationHost applicationHost)
     {
         serviceCollection.AddSingleton<IXtreamClient, XtreamClient>();
+        serviceCollection.AddSingleton<RetryHandler>();
+        serviceCollection.AddSingleton<FailureTrackingService>();
         serviceCollection.AddSingleton<ILiveTvService, LiveTvService>();
         serviceCollection.AddSingleton<IChannel, CatchupChannel>();
         serviceCollection.AddSingleton<IChannel, SeriesChannel>();